Retornado para classe única para testes

+++ Ferramenta/core/adaptees/SateliteImagery/cbers_imagery_service.py
@@ -4,6 +4,8 @@ import os
 import urllib
 from concurrent import futures
 from concurrent.futures.thread import ThreadPoolExecutor
+
+import arcpy
 import pandas as pd
 import requests
 from datetime import datetime
@@ -28,6 +30,10 @@ class CBERSImageryService(BaseImageAcquisitionService):
         super().__init__(*args, **kwargs)
         self.tiles_layer = Feature(path=self.base_gbd.full_path, name=self.__tiles_layer_name)
         self.images_database = Database(path=os.path.dirname(self.images_folder), name='CBERS_IMAGES')
+
+    def compose_image(self, files, download_folder: str):
+        out_img = f"{download_folder}\\compose.tif"
+        arcpy.management.CompositeBands(';'.join(files), out_img)
 
     def __get_images_metadata(self, area: [], initial_date: datetime, final_date: datetime) -> pd.DataFrame:
         logging.debug(f'Coordenadas da área: {area}\n'
@@ -79,45 +85,40 @@ class CBERSImageryService(BaseImageAcquisitionService):
         metadata = self.__get_images_metadata(area, initial_date, final_date)
         with ThreadPoolExecutor(max_workers=5) as threads:
             print('Iniciando downloads')
+            files = []
+            thread_list = []
             for index, row in metadata.iterrows():
-                pan_thread = threads.submit(self.__download_worker, row['id'], row['pan_url'], download_folder, 'pan')
-                red_thread = threads.submit(self.__download_worker, row['id'], row['red_url'], download_folder, 'red')
-                green_thread = threads.submit(self.__download_worker, row['id'], row['green_url'], download_folder, 'green')
-                blue_thread = threads.submit(self.__download_worker, row['id'], row['blue_url'], download_folder, 'blue')
-                nir_thread = threads.submit(self.__download_worker, row['id'], row['nir_url'], download_folder, 'nir')
-                futures.wait([pan_thread, red_thread, green_thread, blue_thread, nir_thread])
-                if pan_thread.exception() is not None:
-                    self.__erase_image(row['id'])
-                    raise Exception(f'Falha ao baixar a imagem PAN', str(pan_thread.exception()))
-                if red_thread.exception() is not None:
-                    self.__erase_image(row['id'])
-                    raise Exception(f'Falha ao baixar a imagem RED', str(red_thread.exception()))
-                if green_thread.exception() is not None:
-                    self.__erase_image(row['id'])
-                    raise Exception(f'Falha ao baixar a imagem GREEN', str(green_thread.exception()))
-                if blue_thread.exception() is not None:
-                    self.__erase_image(row['id'])
-                    raise Exception(f'Falha ao baixar a imagem BLUE', str(blue_thread.exception()))
-                if nir_thread.exception() is not None:
-                    self.__erase_image(row['id'])
-                    raise Exception(f'Falha ao baixar a imagem NIR', str(nir_thread.exception()))
+                filepath = f"{download_folder}\\p_{row['id']}.tif"
+                files.append(filepath)
+                thread_list.append(threads.submit(self.__download_worker, row['id'], row['pan_url'], filepath))
+                filepath = f"{download_folder}\\r_{row['id']}.tif"
+                files.append(filepath)
+                thread_list.append(threads.submit(self.__download_worker, row['id'], row['red_url'], filepath))
+                filepath = f"{download_folder}\\g_{row['id']}.tif"
+                files.append(filepath)
+                thread_list.append(threads.submit(self.__download_worker, row['id'], row['green_url'], filepath))
+                filepath = f"{download_folder}\\b_{row['id']}.tif"
+                files.append(filepath)
+                thread_list.append(threads.submit(self.__download_worker, row['id'], row['blue_url'], filepath))
+                filepath = f"{download_folder}\\n_{row['id']}.tif"
+                files.append(filepath)
+                thread_list.append(threads.submit(self.__download_worker, row['id'], row['nir_url'], filepath))
+                futures.wait(thread_list)
+                for thread in thread_list:
+                    if thread.exception() is not None:
+                        self.__erase_image(row['id'])
+                        raise Exception(f'Falha ao baixar a imagem.', str(thread.exception()))
             print('Downloads finalizados')
+            return files
 
-    def __download_worker(self, id: str, url: str, folder: str, image_type: str) -> None:
-        filepath = f"{folder}\\{id}.tif"
-        if image_type == 'pan':
-            filepath = f"{folder}\\p_{id}.tif"
-        elif image_type == 'red':
-            filepath = f"{folder}\\r_{id}.tif"
-        elif image_type == 'green':
-            filepath = f"{folder}\\g_{id}.tif"
-        elif image_type == 'blue':
-            filepath = f"{folder}\\b_{id}.tif"
-        elif image_type == 'nir':
-            filepath = f"{folder}\\n_{id}.tif"
-        logging.debug(f'Baixando imagem {id} => {filepath}')
-        print(f'Baixando imagem {id} => {filepath}')
-        urllib.request.urlretrieve(url, filepath)
+    def __download_worker(self, id: str, url: str, filepath: str) -> None:
+        if os.path.exists(filepath):
+            logging.debug(f'Arquivo de imagem {filepath} já existe')
+            print(f'Arquivo de imagem {filepath} já existe')
+        else:
+            logging.debug(f'Baixando imagem {id} => {filepath}')
+            print(f'Baixando imagem {id} => {filepath}')
+            urllib.request.urlretrieve(url, filepath)
 
     def __erase_image(self, folder: str, id: str) -> None:
         try:
